fix link hovering causing increased width
Fix the hover state of links causing an increase in width, causing the
lines of text in the paragraph to move lines. When a link is hovered,
the theme by default changes its font weight to bold. The increase in
font weight causes the width of the element's text to also increase,
causing words to shift down a line. By instead unsetting the font width,
a text shadow of 1 pixel can be set to emulate the effect of bolding
without actually increasing the font weight.

This fixes text like so (note the newlines):

```
Mantaro is a simple but powerful Discord bot used to enhance your Discord
experience. It includes many fun and useful features such as moderation,
music, games and currency. It’s currently in over 500,000 servers serving
over 35 million users!
```

from changing to this on hover:

```
Mantaro is a simple but powerful Discord bot used to enhance your
Discord experience. It includes many fun and useful features such as
moderation, music, games and currency. It’s currently in over 500,000
servers serving over 35 million users!
```

Signed-off-by: Vivian Hellyer <vivian@hellyer.dev>